chore(core): extract packages/core/ — first cut (#36)
New packages/core/src/ holds the vscode-API-free shared modules:
- themes/ (25 palettes + helpers)
- markdown/tokens.ts (tokenize + inlineToText)
- markdown/renderer.ts — NEW consolidated renderMarkdown +
  applyMermaidPlaceholders that both src/webview/main.ts and
  apps/electron/renderer/renderer.ts now use (kills the ~30 LOC
  mermaid-placeholder duplication that #13 called out)
- secrets/scanner.ts (token regex set)
- semver/compare.ts (compareSemver + parseSemver)

Previous import sites stay green via thin re-export shims at
the original locations — zero call-site churn, tests still
71/71 passing, Electron renderer bundles unchanged at 7.1MB.

tsconfig.json widened: rootDir = "."; include both src/ and
packages/core/src/. package.json#main updated to match the new
out/src/extension.js layout.

What's deferred (documented in packages/core/README.md):
notes/warehouse/publish/slideshow/MCP-install all touch vscode.*
and need interface abstractions before they can land in core.
Electron's IPC bridge currently re-implements small subsets that
core could absorb in a future ship.
