+++ packages/core/src/secrets/index.ts
@@ -1,0 +1,1 @@
+export * from './scanner';
