+++ packages/core/src/semver/index.ts
@@ -1,0 +1,1 @@
+export * from './compare';
