+++ tsconfig.json
@@ -4,7 +4,7 @@
     "target": "ES2022",
     "lib": ["ES2022"],
     "outDir": "out",
-    "rootDir": "src",
+    "rootDir": ".",
     "sourceMap": true,
     "strict": true,
     "moduleResolution": "Node16",
@@ -13,6 +13,6 @@
     "forceConsistentCasingInFileNames": true,
     "resolveJsonModule": true
   },
-  "include": ["src/**/*"],
+  "include": ["src/**/*", "packages/core/src/**/*"],
   "exclude": ["node_modules", "out", "src/webview/**/*"]
 }
